requirements.txt
discord.py >= 1.7.3
cohere >= 1.3.2